+++ src/logChecker/__init__.py
@@ -1,2 +1,2 @@
-__version__ = "4.5.7"
+__version__ = "4.5.8"
 __author__ = 'Lucas Aimaretto, Beatriz Bonafe'